requirements.txt
fastapi>=0.116
uvicorn[standard]>=0.35
pydantic>=2.11
python-dotenv>=1.1
fastapi-users[sqlalchemy]>=14.0
httpx-oauth>=0.15
python-multipart>=0.0.20
pydantic-settings>=2.10
alembic>=1.15
asyncpg>=0.30
psycopg2-binary>=2.9
gunicorn>=21.2,<22.0
stripe>=8.0
